requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
base58==2.1.1
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cytoolz==0.12.1
dlib==19.24.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.92.0
frozenlist==1.3.3
h11==0.14.0
hexbytes==0.3.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
multiaddr==0.0.9
multidict==6.0.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
netaddr==0.8.0
numpy==1.24.2
orjson==3.8.6
parsimonious==0.8.1
Pillow==9.4.0
protobuf==3.19.5
pycodestyle==2.10.0
pycryptodome==3.17
pydantic==1.10.5
PyMySQL==1.0.2
pyrsistent==0.19.3
python-dotenv==0.21.1
python-multipart==0.0.5
pywin32==305
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rlp==2.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
varint==1.0.2
watchfiles==0.18.1
web3==5.31.3
websockets==9.1
yarl==1.8.2
